vapi-ai-caller/requirements.txt
pandas==2.1.0
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
groq==0.4.0
vapi_python